Now properly builds and configures cmake so others can find it. Includes installation and uninstallation.
Merge commit 'd786f39ab1bc406c82c5db401fbe450ca6073a05'

# Conflicts:
#	src/CMakeLists.txt
