+++ include/icpcuda/Cuda/internal.h
