+++ include/icpcuda/ICPSlowdometry.h
@@ -9,7 +9,7 @@
 #define ICPSLOWDOMETRY_H_
 
 #include "Cuda/internal.h"
-#include "OdometryProvider.h"
+#include "icpcuda/OdometryProvider.h"
 
 #include <opencv2/opencv.hpp>
 #include <vector>
