+++ src/Cuda/containers/device_memory.cpp
@@ -34,8 +34,8 @@
  *  Author: Anatoly Baskeheev, Itseez Ltd, (myname.mysurname@mycompany.com)
  */
 
-#include "device_memory.hpp"
-#include "safe_call.hpp"
+#include "icpcuda/Cuda/containers/device_memory.hpp"
+#include "icpcuda/Cuda/containers/safe_call.hpp"
 
 #include "cuda_runtime_api.h"
 #include "assert.h"
