+++ src/Cuda/containers/initialization.cpp
@@ -34,9 +34,9 @@
 *  Author: Anatoly Baskeheev, Itseez Ltd, (myname.mysurname@mycompany.com)
 */
 
-#include "initialization.hpp"
+#include "icpcuda/Cuda/containers/initialization.hpp"
+#include "icpcuda/Cuda/containers/safe_call.hpp"
 #include "cuda.h"
-#include "safe_call.hpp"
 #include <stdio.h>
 #include <cassert>
 #include <iostream>
