+++ src/ICP.cpp.orig
@@ -1,5 +1,5 @@
-#include "icpcuda/ICPOdometry.h"
-#include "icpcuda/ICPSlowdometry.h"
+#include "ICPOdometry.h"
+#include "ICPSlowdometry.h"
 
 #include <iomanip>
 #include <fstream>
@@ -84,8 +84,13 @@ void outputFreiburg(const std::string filename, const int64_t & timestamp, const
 
 int main(int argc, char * argv[])
 {
-    namespace ch = std::chrono;
+<<<<<<< HEAD
     assert((argc == 2 || argc == 3) && "Please supply the depth.txt dir as the first argument");
+=======
+    namespace ch = std::chrono;
+
+    assert((argc == 2 || argc == 3) && "Please supply the association file dir as the first argument");
+>>>>>>> 6638047f3c717f5ae2091664006b0f43850dfc18
 
     directory.append(argv[1]);
 
@@ -165,11 +170,25 @@ int main(int argc, char * argv[])
                         Eigen::Vector3f trans = currPoseFast.topRightCorner(3, 1);
                         Eigen::Matrix<float, 3, 3, Eigen::RowMajor> rot = currPoseFast.topLeftCorner(3, 3);
 
+<<<<<<< HEAD
+                        boost::posix_time::ptime time = boost::posix_time::microsec_clock::local_time();
+                        boost::posix_time::time_duration duration1(time.time_of_day());
+                        unsigned long long int tick = duration1.total_microseconds();
+
+                        icpOdom.getIncrementalTransformation(trans, rot, threads, blocks);
+
+                        time = boost::posix_time::microsec_clock::local_time();
+                        boost::posix_time::time_duration duration2(time.time_of_day());
+                        unsigned long long int tock = duration2.total_microseconds();
+
+                        meanFast = (float(count) * meanFast + (tock - tick) / 1000.0f) / float(count + 1);
+=======
                         auto tick = ch::high_resolution_clock::now();
                         icpOdom.getIncrementalTransformation(trans, rot, threads, blocks);
                         auto tock = ch::high_resolution_clock::now();
 
                         meanFast = (float(count) * meanFast + ch::duration_cast< ch::milliseconds >(tock-tick).count()) / float(count + 1);
+>>>>>>> 6638047f3c717f5ae2091664006b0f43850dfc18
                         count++;
                     }
 
@@ -206,9 +225,21 @@ int main(int argc, char * argv[])
         Eigen::Vector3f trans = currPoseFast.topRightCorner(3, 1);
         Eigen::Matrix<float, 3, 3, Eigen::RowMajor> rot = currPoseFast.topLeftCorner(3, 3);
 
+<<<<<<< HEAD
+        boost::posix_time::ptime time = boost::posix_time::microsec_clock::local_time();
+        boost::posix_time::time_duration duration1(time.time_of_day());
+        unsigned long long int tick = duration1.total_microseconds();
+
+        icpOdom.getIncrementalTransformation(trans, rot, threads, blocks);
+
+        time = boost::posix_time::microsec_clock::local_time();
+        boost::posix_time::time_duration duration2(time.time_of_day());
+        unsigned long long int tock = duration2.total_microseconds();
+=======
         auto tick_fast = ch::high_resolution_clock::now();
         icpOdom.getIncrementalTransformation(trans, rot, threads, blocks);
         auto tock_fast = ch::high_resolution_clock::now();
+>>>>>>> 6638047f3c717f5ae2091664006b0f43850dfc18
 
         currPoseFast.topLeftCorner(3, 3) = rot;
         currPoseFast.topRightCorner(3, 1) = trans;
@@ -220,15 +251,32 @@ int main(int argc, char * argv[])
         trans = currPoseSlow.topRightCorner(3, 1);
         rot = currPoseSlow.topLeftCorner(3, 3);
 
+<<<<<<< HEAD
+        time = boost::posix_time::microsec_clock::local_time();
+        boost::posix_time::time_duration duration3(time.time_of_day());
+        unsigned long long int ticks = duration3.total_microseconds();
+
+        icpSlowdom.getIncrementalTransformation(trans, rot);
+
+        time = boost::posix_time::microsec_clock::local_time();
+        boost::posix_time::time_duration duration4(time.time_of_day());
+        unsigned long long int tocks = duration4.total_microseconds();
+=======
         auto tick_slow = ch::high_resolution_clock::now();
         icpSlowdom.getIncrementalTransformation(trans, rot);
         auto tock_slow = ch::high_resolution_clock::now();
+>>>>>>> 6638047f3c717f5ae2091664006b0f43850dfc18
 
         currPoseSlow.topLeftCorner(3, 3) = rot;
         currPoseSlow.topRightCorner(3, 1) = trans;
 
+<<<<<<< HEAD
+        meanFast = (float(count) * meanFast + (tock - tick) / 1000.0f) / float(count + 1);
+        meanSlow = (float(count) * meanSlow + (tocks - ticks) / 1000.0f) / float(count + 1);
+=======
         meanFast = (float(count) * meanFast + ch::duration_cast< ch::milliseconds >(tock_fast - tick_fast).count()) / float(count + 1);
         meanSlow = (float(count) * meanSlow + ch::duration_cast< ch::milliseconds >(tock_slow - tick_slow).count()) / float(count + 1);
+>>>>>>> 6638047f3c717f5ae2091664006b0f43850dfc18
         count++;
 
         std::cout << std::setprecision(4) << std::fixed
@@ -248,7 +296,7 @@ int main(int argc, char * argv[])
 
     std::cout << std::endl;
 
-    std::cout << meanSlow / meanFast << " times faster." << std::endl;
+    std::cout << meanSlow / meanFast << " times faster. Fast ICP speed: " << int(1000.f / meanFast) << "Hz" << std::endl;
 
     return 0;
 }
